MyApps-16: Move mixer to viewmodel submodule.
