+++ presentation/viewmodel/src/main/java/xyz/torquato/myapps/presentation/viewmodel/mixer/model/MixerUiState.kt
@@ -1,7 +1,4 @@
-package xyz.torquato.myapps.ui.mixer.model
-
-import xyz.torquato.myapps.domain.impl.mixer.model.Tone
-
+package xyz.torquato.myapps.presentation.viewmodel.mixer.model
 
 data class MixerUiState(
     val isRecording: Boolean,
