+++ presentation/viewmodel/src/main/java/xyz/torquato/myapps/presentation/viewmodel/mixer/model/Tone.kt
@@ -1,0 +1,6 @@
+package xyz.torquato.myapps.presentation.viewmodel.mixer.model
+
+data class Tone (
+    val frequency: Float,
+    val amplitude: Float,
+)
