+++ presentation/viewmodel/src/main/java/xyz/torquato/myapps/presentation/viewmodel/music/MusicController.kt
@@ -1,0 +1,14 @@
+package xyz.torquato.myapps.presentation.viewmodel.music
+
+import androidx.lifecycle.ViewModel
+import xyz.torquato.myapps.domain.impl.mixer.SetChannelStateUseCase
+import javax.inject.Inject
+
+class MusicController @Inject constructor(
+    private val setChannelStateUseCase: SetChannelStateUseCase
+): ViewModel() {
+
+    fun enabledChannel(enable: Boolean) {
+        setChannelStateUseCase(enable)
+    }
+}
